Default fonts stay rooted, the glyph atlas uploads only its dirty rectangles, and post-process proxies are released on the render thread
- Default font/sprite caches are TStrongObjectPtr and retry after a failed load; FT_Library leaks on every failed InitFreeType are closed; culture changes keep the old face until the new bytes are in hand; editor-side inline fonts feed the async rasterizer or fall back to sync with a warning.
- Screen-space view parameters are copied to a render-thread struct through a render command instead of being read across threads; the fullscreen quad resources live in one TU.
- Per-slice dirty rectangles are uploaded via pooled staging textures and CopyTexture(DestSliceIndex); more than half a slice falls back to the whole-slice lock.
- Post-process render proxies are thread-safe shared pointers held by the visual, the mesh section and in-flight commands; the last reference always drops on the render thread. Section updates capture what they need on the game thread.
- The 32-bit index typedef compiles again, the text painter stops before it overflows the index type, gizmo vertex rebuilds init the right buffer, and kerning/line-height/vertical-offset caches stop truncating to integers.

+++ Source/DreamGUI/Public/Core/DreamUIFontData_DistanceField.h
@@ -92,12 +92,12 @@ private:
 	 */
 	UPROPERTY(EditAnywhere, Category = "DreamGUI", meta = (UIMin = "0.0", UIMax = "0.2"))
 		float BoldRatio = 0.04f;
-	/** -1 means not set yet. */
+	/** -1 means not set yet. Transient, so widening it costs no serialized data. */
 	UPROPERTY(VisibleAnywhere, Transient, Category = "DreamGUI", Transient)
-		int LineHeight = -1;
-	/** -1 means not set yet. */
+		float LineHeight = -1;
+	/** -1 means not set yet. Transient, so widening it costs no serialized data. */
 	UPROPERTY(VisibleAnywhere, Transient, Category = "DreamGUI", Transient)
-		int VerticalOffset = -1;
+		float VerticalOffset = -1;
 	/** Ascent and descent at SampleFontSize; negative means not cached yet. */
 	float CachedAscent = -1.0f;
 	float CachedDescent = -1.0f;
@@ -140,7 +140,9 @@ public:
 protected:
 	float OneDivideFontSize = 1.0f; float ExpandMeshSize = 0;
 	TMap<FDreamUIDistanceFieldCharKey, FDreamUICharData> CharDataMap;
-	TMap<FDreamUIDistanceFieldFontKerningPair, int16> KerningPairsMap;
+	// Kerning is a fraction of a pixel at SampleFontSize as often as not; int16 rounded every one of
+	// those to zero, and the first (uncached) call disagreed with every later one. Not serialized.
+	TMap<FDreamUIDistanceFieldFontKerningPair, float> KerningPairsMap;
 	virtual UTexture2DArray* CreateFontTexture(int InTextureSize, int InSliceCount)override;
 	virtual void ApplyPackingAtlasTextureExpand(UTexture2D* newTexture, int newTextureSize)override;
